Vastly improve performance of day 5
Instead of returning a big Vec of all the results to then sort and pick
the last (throwing away almost all of the vec in the process), we just
keep a record of the lowest seen value and update this if we see a lower
one. Drops the runtime for p2 on my laptop by approx 50%, presumably due
to a lot fewer allocations/vec extensions, and not having to needlessly
sort a massive array at the end.